# :bug: improve decompile performance