This improves decompilation performance significantly. Also adds more tracing spans so we get more visibility into decomp perf in future.

Before:

* Full analysis via kantra in source + deps mode
    _InProgress_

* Decompile performance (41 sec per file, 22451 sec total)

    ![Screenshot_20240308_133015](https://github.com/konveyor/analyzer-lsp/assets/9839757/f4e282c0-bce2-4930-a828-204eb3c98a4d)


After:

* Full analysis via kantra in source + deps mode (improved by )
     ```
     real	10m21.651s
     user	0m1.978s
     sys	0m2.554s
     ```
     
* Decompile performance (3.1 sec on per file, 1683 sec total, improved by ~93%)

    ![Screenshot_20240308_133054](https://github.com/konveyor/analyzer-lsp/assets/9839757/22d9b883-baef-4aef-8f07-36f4e18d693b)

    
There's one more issue we need to fix. We are decompiling for dependency CLI too. That's because we use the same _Init()_ function for analyzer and dep cli. We either need to combine these two commands, or have an alternate _InitDep()_ function interface on providers that does a light weight init of providers. Captured in https://github.com/konveyor/analyzer-lsp/issues/529